+++ packages/yandex-connector/src/yandex_connector/server.py
@@ -71,7 +71,7 @@ from yandex_connector.settings import get_settings
 
 _settings = get_settings()
 
-SERVER_VERSION = "1.7.0"
+SERVER_VERSION = "1.8.0"
 SERVER_STARTED_AT = datetime.datetime.now(datetime.UTC).isoformat().replace("+00:00", "Z")
 
 SITE_BASE = "https://market.yandex.ru"